refactor: move SQL models and add dataclass entities
