+++ app/domain/models.py
@@ -1,15 +1,5 @@
-from datetime import datetime
-from ..extensions import db
+"""Backward compatibility module for SQLAlchemy models."""
 
+from ..adapters.sqlalchemy.models import User, Product
 
-class User(db.Model):
-    id = db.Column(db.Integer, primary_key=True)
-    username = db.Column(db.String(80), unique=True, nullable=False)
-    password_hash = db.Column(db.String(128), nullable=False)
-
-
-class Product(db.Model):
-    id = db.Column(db.Integer, primary_key=True)
-    name = db.Column(db.String(120), nullable=False)
-    price = db.Column(db.Float, nullable=False)
-    created_at = db.Column(db.DateTime, default=datetime.utcnow)
+__all__ = ["User", "Product"]
