+++ app/services/product_service.py
@@ -1,4 +1,4 @@
-from ..domain.models import Product
+from ..adapters.sqlalchemy.models import Product
 from ..domain.repositories import ProductRepository
